0-async_generator.py

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -6,7 +6,7 @@ import random
 from typing import AsyncGenerator
 
 
-async def async_generator() -> AsyncGenerator[float, None]:
+async def async_generator() -> AsyncGenerator[int, None]:
     """Coroutine that loop 10 times, each time asynchronously wait 1 second,
     then yield a random number between 0 and 10. Use the random module."""
     for _ in range(10):
